Extract classes from `ConfigurationCacheReport.kt`
For smaller, more focused diffs.
