+++ subprojects/configuration-cache/src/main/kotlin/org/gradle/configurationcache/problems/StackTraceExtractor.kt
@@ -1,0 +1,38 @@
+/*
+ * Copyright 2021 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.configurationcache.problems
+
+import java.io.PrintWriter
+import java.io.StringWriter
+
+
+internal
+class StackTraceExtractor {
+
+    private
+    val stringWriter = StringWriter()
+
+    private
+    val printWriter = PrintWriter(stringWriter)
+
+    fun stackTraceStringFor(error: Throwable): String {
+        error.printStackTrace(printWriter)
+        return stringWriter.toString().also {
+            stringWriter.buffer.setLength(0)
+        }
+    }
+}
